Move files around to standard bin/lib folders
