+++ lib/ast.js
